Organize tests folder structure
